[refactor] #94 navigateVideoDetail 인자 수정에 따른 코드 리팩토링

+++ feature/mypage/src/main/java/com/record/mypage/navigation/MypageNavigation.kt
@@ -29,7 +29,7 @@ fun NavGraphBuilder.mypageNavGraph(
     navigateToSetting: () -> Unit,
     navigateToFollowing: () -> Unit,
     navigateToFollower: () -> Unit,
-    navigateToVideo: (VideoType, Long) -> Unit,
+    navigateToVideo: (VideoType, Int) -> Unit,
 ) {
     composable(route = MypageRoute.route) {
         MypageRoute(
